# api/openapi/openapi.yaml
# Generated with protoc-gen-openapi
# https://github.com/google/gnostic/tree/master/cmd/protoc-gen-openapi

openapi: 3.0.3
info:
    title: UserCenter API
    version: 0.0.1
paths:
    /v1/usercenter/login:
        post:
            tags:
                - UserCenter
            description: Login 登录
            operationId: UserCenter_Login
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/usercenter.v1.LoginRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/usercenter.v1.LoginReply'
    /v1/usercenter/logout:
        post:
            tags:
                - UserCenter
            description: Logout 登出
            operationId: UserCenter_Logout
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/usercenter.v1.LogoutRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content: {}
    /v1/users:
        post:
            tags:
                - UserCenter
            description: CreateUser
            operationId: UserCenter_CreateUser
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/usercenter.v1.CreateUserRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/usercenter.v1.UserReply'
components:
    schemas:
        usercenter.v1.CreateUserRequest:
            type: object
            properties:
                username:
                    type: string
                nickname:
                    type: string
                password:
                    type: string
                email:
                    type: string
                phone:
                    type: string
        usercenter.v1.LoginReply:
            type: object
            properties:
                refreshToken:
                    type: string
                accessToken:
                    type: string
                type:
                    type: string
                expiresAt:
                    type: integer
                    format: int64
            description: LoginReply 登录响应
        usercenter.v1.LoginRequest:
            type: object
            properties:
                username:
                    type: string
                password:
                    type: string
            description: LoginRequest 登录请求
        usercenter.v1.LogoutRequest:
            type: object
            properties: {}
        usercenter.v1.UserReply:
            type: object
            properties:
                userID:
                    type: string
                username:
                    type: string
                nickname:
                    type: string
                password:
                    type: string
                email:
                    type: string
                phone:
                    type: string
                secrets:
                    type: integer
                    format: int64
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
tags:
    - name: UserCenter